refactor(LUKE-132): move the live session machinery behind @sidecar/voice/live-session
The transport-neutral live session service and its parts (the append
channel, the proactive queue, the roster context, the trace vocabulary, the
LiveBrain and LiveRecord doors, and the graceful close) move from
packages/host/src/voice into @sidecar/voice behind a live-session door, with
their tests. The host composes the service from its new home unchanged; the
desktop's Conversation writer and the BrainAgent adapter stay in the host as
the door's two implementations there, and the ws socket seam stays in the
host too, so @sidecar/voice names no socket library and a web function can
bundle the door.

No behaviour changes: every moved body is unchanged, only imports and the
file split move.
