+++ packages/host/src/voice/conversation-live-record.test.ts
@@ -2,7 +2,7 @@ import assert from "node:assert/strict";
 import { MAIN_SESSION_KEY, type SessionKey } from "@sidecar/runtime/vocabulary";
 import { CONVERSATION_ENTRY_KIND, type ConversationEntry } from "@sidecar/session";
 import { test } from "vitest";
-import { conversationLiveRecord } from "./live-record.js";
+import { conversationLiveRecord } from "./conversation-live-record.js";
 
 function writer() {
   const written: { entry: ConversationEntry; recordedAt: number; sessionKey: SessionKey }[] = [];
