+++ packages/host/src/voice/socket-over-ws.test.ts
@@ -2,18 +2,10 @@ import assert from "node:assert/strict";
 import { once } from "node:events";
 import http from "node:http";
 import type { AddressInfo } from "node:net";
-import { LIVE_CLIENT_EVENT, LIVE_CLOSE_REASON, type LiveServerEvent } from "@sidecar/live";
-import { drainMicrotasks, FakeClock } from "@sidecar/runtime/testing";
-import { SOCKET_OPEN_FAULT, sidebandOverSocket, socketOpened } from "@sidecar/voice";
-import { FakeLiveSocket } from "@sidecar/voice/testing";
+import { SOCKET_OPEN_FAULT, socketOpened } from "@sidecar/voice";
 import { test } from "vitest";
 import { WebSocketServer } from "ws";
-import {
-  closeGracefully,
-  openSocketOverWs,
-  SIDEBAND_CLOSE_OUTCOME,
-  SIDEBAND_CLOSE_TIMEOUT_MS,
-} from "./live-sideband.js";
+import { openSocketOverWs } from "./socket-over-ws.js";
 
 /** A local upgrade endpoint that admits a bearer and refuses everything else with a status. */
 async function server(options: { bearer: string; refuseWith: number }) {
@@ -92,69 +84,4 @@ test("a refused upgrade answers its status, and nothing listening answers a netw
   assert.equal(socketOpened(unreachable), false);
   if (socketOpened(unreachable)) return;
   assert.equal(unreachable.fault, SOCKET_OPEN_FAULT.NETWORK);
-});
-
-function closedEvent(reason: string, seconds: number) {
-  return {
-    type: "session.closed",
-    event_id: "closed",
-    reason,
-    usage: { seconds },
-  };
-}
-
-test("a graceful close registers the closed listener, sends session.close, and holds the socket until the final event", async () => {
-  const socket = new FakeLiveSocket();
-  const sideband = sidebandOverSocket(socket);
-  const clock = new FakeClock();
-  const heard: LiveServerEvent[] = [];
-  sideband.onEvent((event) => heard.push(event));
-  const closing = closeGracefully(sideband, {
-    eventId: "close-1",
-    schedule: clock.schedule,
-    cancel: clock.cancel,
-  });
-  await drainMicrotasks();
-  assert.deepEqual(
-    socket.sent.map((frame) => JSON.parse(frame)),
-    [{ type: LIVE_CLIENT_EVENT.CLOSE, event_id: "close-1" }],
-  );
-  assert.equal(socket.closedByClient, false);
-  assert.deepEqual(clock.delays, [SIDEBAND_CLOSE_TIMEOUT_MS]);
-  socket.receive(closedEvent(LIVE_CLOSE_REASON.CLOSE_REQUESTED, 61));
-  const result = await closing;
-  assert.equal(result.outcome, SIDEBAND_CLOSE_OUTCOME.CLOSED);
-  if (result.outcome !== SIDEBAND_CLOSE_OUTCOME.CLOSED) return;
-  assert.deepEqual(
-    [result.closed.reason, result.closed.usage.seconds],
-    [LIVE_CLOSE_REASON.CLOSE_REQUESTED, 61],
-  );
-  assert.equal(socket.closedByClient, true);
-  assert.equal(clock.armed(), 0);
-});
-
-test("a socket that ends first leaves the close unconfirmed, and silence gives up at the timeout", async () => {
-  const lost = new FakeLiveSocket();
-  const clock = new FakeClock();
-  const losing = closeGracefully(sidebandOverSocket(lost), {
-    eventId: "c",
-    schedule: clock.schedule,
-    cancel: clock.cancel,
-  });
-  lost.closeFromServer({ code: 1006 });
-  assert.deepEqual(await losing, {
-    outcome: SIDEBAND_CLOSE_OUTCOME.CONNECTION_LOST,
-    close: { code: 1006 },
-  });
-
-  const silent = new FakeLiveSocket();
-  const timing = closeGracefully(sidebandOverSocket(silent), {
-    eventId: "c",
-    schedule: clock.schedule,
-    cancel: clock.cancel,
-    timeoutMs: 50,
-  });
-  await clock.advance(clock.now + 50);
-  assert.deepEqual(await timing, { outcome: SIDEBAND_CLOSE_OUTCOME.TIMED_OUT });
-  assert.equal(silent.closedByClient, true);
 });
