+++ packages/voice/src/live-session/index.ts
@@ -1,0 +1,11 @@
+export {
+  LIVE_BRAIN_RUN_END,
+  LIVE_BRAIN_RUN_EVENT,
+  LIVE_BRAIN_SUBMISSION,
+  type LiveBrain,
+  type LiveBrainRunEnd,
+  type LiveBrainRunEvent,
+} from "./live-brain.js";
+export type { LiveRecord } from "./live-record.js";
+export { LiveSessionService } from "./live-session-service.js";
+export type { BeatKind } from "./proactive-queue.js";
